Split initial index.ts file into app.ts and index.ts. This was done in order to be able to import the app.ts file into supertest
