+++ .env
@@ -1,2 +1,3 @@
 DEV_POSTGRES_URL=postgresql://postgres:example@localhost:5436/neighborhood-dev
 TEST_POSTGRES_URL=postgresql://postgres:example@localhost:25432/neighborhood-test
+PORT=3000
